requirements.txt
streamlit
pillow
numpy
pandas
